Fix [V2]: Elastic Agent Install is broken.
